some more code cleanup, but i guess nothing speed improving since the java compiler should be intelligent enough to optimize it at compile time. but for the look of the code it's fine too :)
git-svn-id: http://svn.l2jfree.com/l2jfree@4897 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/l2jfree-core/src/main/java/com/l2jfree/gameserver/model/zone/L2Zone.java
@@ -407,7 +407,7 @@ public abstract class L2Zone
 		{
 			sum += sh.getMiddleX();
 		}
-		return (int)(sum / _shapes.length);
+		return (sum / _shapes.length);
 	}
 
 	public int getMiddleY()
@@ -423,7 +423,7 @@ public abstract class L2Zone
 		{
 			sum += sh.getMiddleY();
 		}
-		return (int)(sum / _shapes.length);
+		return (sum / _shapes.length);
 	}
 
 	public boolean intersectsRectangle(int ax, int bx, int ay, int by)
